feat: Add default example on import (#3041)
