+++ packages/api-client/src/store/request-example.ts
@@ -1,15 +1,10 @@
 import type { StoreContext } from '@/store/store-context'
 import {
-  type ExampleRequestBody,
   type Request,
   type RequestExample,
-  type RequestExampleParameter,
-  type RequestParameter,
-  requestExampleParametersSchema,
-  requestExampleSchema,
+  createExampleFromRequest,
 } from '@scalar/oas-utils/entities/spec'
 import { iterateTitle } from '@scalar/oas-utils/helpers'
-import { getRequestBodyFromOperation } from '@scalar/oas-utils/spec-getters'
 import { mutationFactory } from '@scalar/object-utils/mutator-record'
 import { reactive } from 'vue'
 
@@ -83,116 +78,4 @@ export function extendedExampleDataFactory({
     addRequestExample,
     deleteRequestExample,
   }
-}
-
-// ---------------------------------------------------------------------------
-// Example Helpers
-
-/** Create new instance parameter from a request parameter */
-export function createParamInstance(param: RequestParameter) {
-  const schema = param.schema as any
-
-  /**
-   * TODO:
-   * - Need better value defaulting here
-   * - Need to handle non-string parameters much better
-   * - Need to handle unions/array values for schema
-   */
-  const value = String(schema.default ?? schema?.examples?.[0] ?? '')
-
-  return requestExampleParametersSchema.parse({
-    ...schema,
-    key: param.name,
-    value,
-    description: param.description,
-    required: param.required,
-    /** Initialized all required properties to enabled */
-    enabled: !!param.required,
-  })
-}
-
-/**
- * Create new request example from a request
- * Iterates the name of the example if provided
- */
-export function createExampleFromRequest(
-  request: Request,
-  name: string,
-): RequestExample {
-  // ---------------------------------------------------------------------------
-  // Populate all parameters with an example value
-  const parameters: Record<
-    'path' | 'cookie' | 'header' | 'query',
-    RequestExampleParameter[]
-  > = {
-    path: [],
-    query: [],
-    cookie: [],
-    header: [],
-  }
-
-  // Populated the separated params
-  request.parameters?.forEach((p) =>
-    parameters[p.in].push(createParamInstance(p)),
-  )
-
-  // ---------------------------------------------------------------------------
-  // Handle request body defaulting for various content type encodings
-  const body: ExampleRequestBody = {
-    activeBody: 'raw',
-    raw: {
-      encoding: 'json',
-      value: '',
-    },
-  }
-
-  if (request.requestBody) {
-    const requestBody = getRequestBodyFromOperation({
-      path: request.path,
-      information: {
-        requestBody: request.requestBody,
-      },
-    })
-
-    if (requestBody?.body?.mimeType === 'application/json') {
-      body.activeBody = 'raw'
-      body.raw = {
-        encoding: 'json',
-        value: requestBody.body.text ?? JSON.stringify({}),
-      }
-    }
-
-    if (requestBody?.body?.mimeType === 'application/xml') {
-      body.activeBody = 'raw'
-      body.raw = {
-        encoding: 'xml',
-        value: requestBody.body.text ?? '',
-      }
-    }
-
-    /**
-     *  TODO: Are we loading example files from somewhere based on the spec?
-     *  How are we handling the body values
-     */
-    if (requestBody?.body?.mimeType === 'application/octet-stream') {
-      body.activeBody = 'binary'
-      body.binary = undefined
-    }
-
-    /**
-     * TODO: How are handling form data examples from the spec
-     */
-    if (requestBody?.body?.mimeType === 'application/x-www-form-urlencoded') {
-      body.activeBody = 'formData'
-      body.formData = undefined
-    }
-  }
-  const example = requestExampleSchema.parse({
-    requestUid: request.uid,
-    parameters,
-    name,
-    body,
-  })
-
-  return example
 }
